+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 All notable changes to `php-csv` will be documented in this file
 
+## 0.2.0 - 2018-01-14
+### Added
+- Generators will now return number of lines, if not aborted
+- `$csv->process()` will now return number of lines, if not aborted 
+
 ## 0.1.0 - 2018-01-13
 Initial Release of `php-csv`